Make AlarmUnit remember changes for more efficient UI updates.
